src/main/java/org/araujo/Player.java
package org.araujo;

public class Player implements Cloneable {

    private String name;
    private int level;
    private int score;

    public Player(String name, int level) {
        this.name = name;
        this.level = level;
        this.score = 0;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getLevel() {
        return level;
    }

    public void setLevel(int level) {
        this.level = level;
    }

    public int getScore() {
        return score;
    }

    public void addScore(int score) {
        this.score += score;
    }

    @Override
    public Player clone() throws CloneNotSupportedException {
        return (Player) super.clone();
    }
}
